+++ packages/common-webpack/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [14.0.2](https://github.com/darkadept/thr-addons/compare/v14.0.1...v14.0.2) (2021-05-28)
+
+
+### Features
+
+* added promise-sequential package ([#307](https://github.com/darkadept/thr-addons/issues/307)) ([b4f94b2](https://github.com/darkadept/thr-addons/commit/b4f94b26507428c21dc36b722a61cc280d05b455))
+
+
+
+
+
 # [14.0.0](https://github.com/darkadept/thr-addons/compare/v13.8.2...v14.0.0) (2021-05-11)
 
 **Note:** Version bump only for package @thx/common-webpack