+++ packages/pdfform/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [14.0.2](https://github.com/thr-consulting/thr-addons/compare/v14.0.1...v14.0.2) (2021-05-28)
+
+**Note:** Version bump only for package @thx/pdfform
+
+
+
+
+
+## [14.0.1](https://github.com/thr-consulting/thr-addons/compare/v14.0.0...v14.0.1) (2021-05-21)
+
+**Note:** Version bump only for package @thx/pdfform
+
+
+
+
+
 # [14.0.0](https://github.com/thr-consulting/thr-addons/compare/v13.8.2...v14.0.0) (2021-05-11)